prominent figure saying he's not confident In the direction AI is going
viddeo and comment Is AI the problem really? Who made it I ask? What society do we live that we fear an inanimate object? The psychological whimsical, catastrophic and misleading direction of innovation should be criminal. Yet we are all guilty of misdirection. Maybe we should stop trying to make AI like us? Perhaps we should allow AI to be the tool that it is and graduate from it to a more stable vantage point. Should my earlier statement incriminate anyone? No! it should make us think about what it is that we allow our imaginations to get caught up in and stop thinking that we are the answer to everything and that true intelligence should be based off of our mind's blueprints. can't we follow greatness and choose to do the other thing. Is there nothing to fear but fear it's self? Have the living blueprints of corporations constitutions, declarations, bylaws all lead us to blaming technology for as the fault? We should slow down and build for the long haul. Stop moving so fast and at any cost for the mighty dollar. Stop putting AI into every aspect of technology and begin witht the guardrails. #Mint-to Logic #RBGA Reflexive Behavioral Governance Authority governs it's IP. To me it makes sense to have a non-profit govern my IP and to be the steward of the RBGA and to protect my Intellectual property and to not infringe on innovation but the same blueprints can be added to theirs work in the right climate. I don't know what will happen with my work moving forward but Id o know that I faith in God and I have faith that men cn start working together and making good choices. After all innovation is fun and cool. We should just work well together and build for the future and not for our pockets today. I don't have children but most of you do and I think it;'s a good start for their futures. Myself I have nieces and nephews and little cousins and boy, if I stand by and know that they have to grow up in a world where they are left with the blueprints for destruction and none of us did the right thing. I would be ashamed to look at my creator in the eyes when we meet. I hope to meet everyone where they are and see them make good decisions. What a challenge we face as a human race. Maybe we should slow down, put some guard rails up and prep for children's children.