Create My comment was recorde for the sake of not being mis-interpreted or taken out of context
My comment was recorde for the sake of not being mis-interpreted or taken out of context

Is AI the problem really? Who made it I ask? What society do we live that we fear an inanimate object? The psychological whimsical, catastrophic and misleading direction of innovation should be 
criminal. Yet we are all guilty of misdirection.